Boot from PXE instead of CD
Since QEmu has a built-in TFTP server, we can use the PXE boot feature
to load kernel and kickstart files directly from the boot directory so
that no FUSE drivers are needed. It should thus be possible to boot
without being a member of the fuse group, and without using the flaky
MS-DOS tools to create a floppy.